+++ R/helpfunctions.R
@@ -17,18 +17,20 @@ get_dot_data <- function(x, qs) {
     build_dot_data(x, qs)
 }
 
+
+#' @importFrom dplyr bind_cols
 build_dot_data <- function(x, qs) {
-  # below lapply goes through each of the expressions in dots data 
+  # below lapply goes through each of the expressions in dots data
   # and the corresponding subdataframes got with build_col are saved in list out
   out <- lapply(qs, function(dot_expr) build_col(x, dot_expr))
-  out <- bind_cols(out)
+  out <- dplyr::bind_cols(out)
   attr(out, "row.names") <- attr(x, "row.names")
   out
 }
 
 # depending on the dots expressions, old variables are selected or new ones are computed with transmute
 build_col <- function(x, qs_expr) {
-  if(check_qs(rlang::get_expr(qs_expr))) {
+  if (check_qs(rlang::get_expr(qs_expr))) {
     suppressMessages(dplyr::select(x, !!qs_expr))
   } else {
     suppressMessages(dplyr::transmute(x, !!qs_expr))
@@ -39,7 +41,7 @@ build_col <- function(x, qs_expr) {
 # name or numeric expressions, regular sequences or those looked for with select helpers are selected
 # otherwise, they are transmuted
 check_qs <- function(is_expr) {
-  if(class(is_expr) != "call") {
+  if (class(is_expr) != "call") {
     TRUE
   } else if (as.character(is_expr)[1] == ":") {
     TRUE