+++ tests/testthat/test-frq.R
@@ -98,3 +98,92 @@ test_that("frq", {
 #'   group_by(c12hour) %>%
 #'   frq(nur_pst, min.frq = 1)
 #'
+
+
+
+test_that("frq, string", {
+  dat <- data.frame(
+    x = c("", "", "a", "a", "b"),
+    stringsAsFactors = FALSE
+  )
+
+  expect_equal(
+    frq(dat$x),
+    structure(list(structure(list(
+      val = structure(
+        c(1L, 2L, 3L, NA),
+        .Label = c("", "a", "b"),
+        class = "factor"
+      ),
+      label = c("<none>", "<none>", "<none>", NA),
+      frq = c(2L, 2L, 1L, 0L),
+      raw.prc = c(40, 40, 20, 0),
+      valid.prc = c(40, 40, 20, NA),
+      cum.prc = c(40, 80, 100, NA)),
+      row.names = c(NA, 4L),
+      class = "data.frame",
+      label = "x",
+      vartype = "character",
+      mean = 1.8,
+      sd = 0.83666)),
+      class = c("sjmisc_frq", "list"),
+      print = "txt",
+      encoding = "UTF-8"
+    ),
+    tolerance = 1e-4
+  )
+
+  expect_equal(
+    frq(dat[["x"]]),
+    structure(list(structure(list(
+      val = structure(
+        c(1L, 2L, 3L, NA),
+        .Label = c("", "a", "b"),
+        class = "factor"
+      ),
+      label = c("<none>", "<none>", "<none>", NA),
+      frq = c(2L, 2L, 1L, 0L),
+      raw.prc = c(40, 40, 20, 0),
+      valid.prc = c(40, 40, 20, NA),
+      cum.prc = c(40, 80, 100, NA)),
+      row.names = c(NA, 4L),
+      class = "data.frame",
+      label = "x",
+      vartype = "character",
+      mean = 1.8,
+      sd = 0.83666)),
+      class = c("sjmisc_frq", "list"),
+      print = "txt",
+      encoding = "UTF-8"
+    ),
+    tolerance = 1e-4
+  )
+
+  expect_equal(
+    frq(dat["x"]),
+    structure(list(structure(list(
+      val = structure(
+        c(1L, 2L, 3L, NA),
+        .Label = c("", "a", "b"),
+        class = "factor"
+      ),
+      label = c("<none>", "<none>", "<none>", NA),
+      frq = c(2L, 2L, 1L, 0L),
+      raw.prc = c(40, 40, 20, 0),
+      valid.prc = c(40, 40, 20, NA),
+      cum.prc = c(40, 80, 100, NA)),
+      row.names = c(NA, 4L),
+      class = "data.frame",
+      label = "x",
+      vartype = "character",
+      mean = 1.8,
+      sd = 0.83666)),
+      class = c("sjmisc_frq", "list"),
+      print = "txt",
+      encoding = "UTF-8"
+    ),
+    tolerance = 1e-4
+  )
+})
+
+